[server.ts 에 mongodb 연결] - server.ts mongode 연결 -> 비동기처리 에러 있어서 수정 필요 - user.ts, article.ts 변수 및 연결 수정
