+++ src/server.ts
@@ -1,5 +1,24 @@
 import app from "./index";
+import mongoose from "mongoose";
+import "dotenv/config";
+
+const { MONGO_USER, MONGO_PASSWORD } = process.env;
+
 const port = process.env.PORT || 3000;
+mongoose.connect(
+  `mongodb://${MONGO_USER}:${MONGO_PASSWORD}:27017/admin`,
+  {
+    dbName: "Balan",
+  },
+  (error) => {
+    if (error) {
+      throw error.message;
+      console.log("몽고디비 연결 에러");
+    } else {
+      console.log("몽고디비 연결 성공!");
+    }
+  }
+);
 
 app.listen(port, () => {
   console.log("server is running");
